fix: vite migrated to use webpack
