+++ Data Fetch (ZKTLS)/tls-notary/README.md
@@ -1,0 +1,8 @@
+# Verifiable Data Fetching with TLS Notary
+
+This section of the project demonstrates a method for securely fetching data using **TLS Notary** for verifiability.
+
+**Key Features:**
+- Ensures the integrity and authenticity of fetched data.
+- Leverages TLS Notary to provide cryptographic proof for the data received.
+- Useful for applications requiring trustless proof of data origin.
