+++ Data Fetch (ZKTLS)/tls-notary/ZK/index.ejs
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>React/Typescrip Example</title>
+</head>
+
+<body>
+<script>
+</script>
+<div id="root"></div>
+</body>
+
+</html>
