+++ Data Fetch (ZKTLS)/tls-notary/ZK/package.json
@@ -1,0 +1,45 @@
+{
+  "name": "ZK",
+  "version": "1.0.0",
+  "description": "",
+  "main": "webpack.js",
+  "scripts": {
+    "dev": "webpack-dev-server --config webpack.js"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "@fortawesome/fontawesome-free": "^6.7.1",
+    "buffer": "^6.0.3",
+    "comlink": "^4.4.2",
+    "crypto-browserify": "^3.12.1",
+    "css-loader": "^7.1.2",
+    "postcss-loader": "^8.1.1",
+    "process": "^0.11.10",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-loader-spinner": "^6.1.6",
+    "stream-browserify": "^3.0.0",
+    "style-loader": "^4.0.0",
+    "tlsn-js": "^0.1.0-alpha.10",
+    "vm-browserify": "^1.1.2"
+  },
+  "devDependencies": {
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.10",
+    "babel-loader": "^9.1.3",
+    "copy-webpack-plugin": "^11.0.0",
+    "html-webpack-plugin": "^5.5.0",
+    "postcss": "^8.4.49",
+    "postcss-preset-env": "^10.1.1",
+    "sass": "^1.82.0",
+    "sass-loader": "^16.0.4",
+    "source-map-loader": "^5.0.0",
+    "tailwindcss": "^3.4.16",
+    "ts-loader": "^9.4.2",
+    "typescript": "^4.9.4",
+    "webpack": "^5.75.0",
+    "webpack-cli": "^4.10.0",
+    "webpack-dev-server": "^4.11.1"
+  }
+}
