+++ Data Fetch (ZKTLS)/tls-notary/ZK/postcss.config.js
@@ -1,0 +1,4 @@
+const tailwindcss = require("tailwindcss");
+module.exports = {
+    plugins: ["postcss-preset-env", tailwindcss],
+};
