+++ Data Fetch (ZKTLS)/tls-notary/ZK/src/app.scss
@@ -1,0 +1,3 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
