+++ Data Fetch (ZKTLS)/tls-notary/ZK/tailwind.config.js
@@ -1,0 +1,12 @@
+/** @type {import('tailwindcss').Config} */
+module.exports = {
+    content: ['./src/**/*.{js,jsx,ts,tsx}'],
+    theme: {
+        extend: {
+            colors: {
+                primary: '#243f5f',
+            },
+        },
+    },
+    plugins: [],
+};
