--- Data Fetch (ZKTLS)/tls-notary/frontend/.vite/deps/_metadata.json
@@ -1,8 +1,0 @@
-{
-  "hash": "6ef5060b",
-  "configHash": "bc84b2c7",
-  "lockfileHash": "e3b0c442",
-  "browserHash": "134e0737",
-  "optimized": {},
-  "chunks": {}
-}
