--- Data Fetch (ZKTLS)/tls-notary/frontend/package.json
@@ -1,62 +1,0 @@
-{
-  "name": "tls-notary",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "tsc -b && vite build",
-    "lint": "eslint .",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "buffer": "^6.0.3",
-    "comlink": "^4.4.2",
-    "crypto-browserify": "^3.12.1",
-    "http-parser-js": "^0.5.10",
-    "process": "^0.11.10",
-    "react": "^19.1.0",
-    "react-dom": "^19.1.0",
-    "stream-browserify": "^3.0.0",
-    "tlsn-js": "^0.1.0-alpha.10.0",
-    "vm-browserify": "^1.1.2"
-  },
-  "devDependencies": {
-    "@babel/preset-env": "^7.27.2",
-    "@babel/preset-react": "^7.27.1",
-    "@babel/preset-typescript": "^7.27.1",
-    "@esbuild-plugins/node-globals-polyfill": "^0.2.3",
-    "@esbuild-plugins/node-modules-polyfill": "^0.2.2",
-    "@eslint/js": "^9.25.0",
-    "@types/node": "^22.15.18",
-    "@types/react": "^19.1.2",
-    "@types/react-dom": "^19.1.2",
-    "@vitejs/plugin-react": "^4.4.1",
-    "autoprefixer": "^10.4.21",
-    "babel-loader": "^10.0.0",
-    "copy-webpack-plugin": "^13.0.0",
-    "css-loader": "^7.1.2",
-    "eslint": "^9.25.0",
-    "eslint-plugin-react-hooks": "^5.2.0",
-    "eslint-plugin-react-refresh": "^0.4.19",
-    "globals": "^16.0.0",
-    "html-webpack-plugin": "^5.6.3",
-    "postcss": "^8.5.3",
-    "postcss-loader": "^8.1.1",
-    "sass": "^1.89.0",
-    "sass-loader": "^16.0.5",
-    "source-map-loader": "^5.0.0",
-    "style-loader": "^4.0.0",
-    "tailwindcss": "^3.4.1",
-    "ts-loader": "^9.5.2",
-    "typescript": "~5.8.3",
-    "typescript-eslint": "^8.30.1",
-    "vite": "^6.3.5",
-    "vite-plugin-comlink": "^5.1.0",
-    "vite-plugin-static-copy": "^3.0.0",
-    "webpack": "^5.99.8",
-    "webpack-cli": "^6.0.1",
-    "webpack-dev-server": "^5.2.1",
-    "zustand": "^4.5.7"
-  }
-}
