--- Data Fetch (ZKTLS)/tls-notary/frontend/src/App.tsx
@@ -1,11 +1,0 @@
-import './App.css'
-import { Home } from './pages/Home';
-
-function App() {
-  return (
-    <Home />
-  );
-}
-
-
-export default App
