--- Data Fetch (ZKTLS)/tls-notary/frontend/src/main.tsx
@@ -1,11 +1,0 @@
-import { StrictMode } from 'react'
-import { createRoot } from 'react-dom/client'
-import './index.css'
-import App from './App.tsx'
-(globalThis as any).__tlsn_wasm_base_url__ = '/tlsn/';
-
-createRoot(document.getElementById('root')!).render(
-  <StrictMode>
-    <App />
-  </StrictMode>,
-)
