--- Data Fetch (ZKTLS)/tls-notary/frontend/tsconfig.json
@@ -1,15 +1,0 @@
-{
-  "files": [],
-  "references": [
-    { "path": "./tsconfig.app.json" },
-    { "path": "./tsconfig.node.json" }
-  ],
-
-  "compilerOptions": {
-    "module": "ESNext",
-    "moduleResolution": "Node",
-    "target": "ESNext",
-    "lib": ["DOM", "ESNext"],
-    "types": ["vite/client"]
-  }
-}
